requirements.txt
agora-token-builder==1.0.0
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
Django==3.2.16
django-agora==0.4
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.6
razorpay==1.3.0
requests==2.28.1
sqlparse==0.4.3
twilio==7.16.0
tzdata==2022.7
urllib3==1.26.13
whitenoise==6.2.0
